+++ src/Main.java
@@ -18,13 +18,18 @@ public class Main extends Application {
 
     private GlyphFont glyphFont;
 
-    private final boolean disableSplashScreen = false;
+    private boolean disableSplashScreen = false;
 
     private Stage splashScreenStage;
     private Stage mainStage;
 
     @Override
     public void start(Stage primaryStage) throws Exception{
+
+        for(String arg : getParameters().getRaw()) {
+            if(arg.equals("no-splash"))
+                disableSplashScreen = true;
+        }
 
         if(disableSplashScreen) {
             // Skip this function, and go straight to loading the main window.
@@ -57,11 +62,6 @@ public class Main extends Application {
         primaryStage.show();
 
         splashScreenViewController.startNameVersionInfoMsg();
-        // Create delay before showing main window
-        /*Timeline timeline = new Timeline(new KeyFrame(
-                Duration.millis(3000),
-                ae -> loadMainWindow()));
-        timeline.play();*/
     }
 
     public void loadMainWindow() {
@@ -100,11 +100,19 @@ public class Main extends Application {
 
         // Make sure the main stage has focus (is in front of all other windows)
         mainStage.toFront();
+
+        // Call event handler in model for app closing
+        mainStage.setOnCloseRequest(event -> {
+            model.handleAppClosing();
+        });
     }
 
 
+    /**
+     * Entry point for application. This calls <code>launch(args)</code> which starts the JavaFX UI.
+     * @param args
+     */
     public static void main(String[] args) {
         launch(args);
-        //LauncherImpl.launchApplication(Main.class, SplashScreenViewController.class, args);
     }
 }