+++ src/ninja/mbedded/ninjaterm/model/Model.java
@@ -30,7 +30,20 @@ public class Model {
 
     }
 
+    /**
+     * Call this to close a terminal.
+     *
+     * This should be called from the UI when the user clicks the "X" button in the tab header,
+     * or when the user clicks close the tab's right-click context menu.
+     *
+     * @param terminalToClose
+     */
     public void closeTerminal(Terminal terminalToClose) {
+
+        // We want to close the COM port connected to this terminal,
+        // if it is open
+        if(terminalToClose.isComPortOpen.get())
+            terminalToClose.closeComPort();
 
         // Emit an event for the UI
         for(CloseTerminalListener closeTerminalListener : closedTerminalListeners) {
@@ -40,6 +53,20 @@ public class Model {
         // The UI should of now removed the terminal Tab and the associated controller
         // Remove the terminal model from the list
         terminals.remove(terminalToClose);
+    }
+
+    /**
+     * This needs to be called when the application is closing. This method
+     * makes sure all the COM ports attached to terminal tabs are closed.
+     */
+    public void handleAppClosing() {
+        // We need to close all COM ports
+        for(Terminal terminal : terminals) {
+            // We want to close the COM port connected to this terminal,
+            // if it is open
+            if(terminal.isComPortOpen.get())
+                terminal.closeComPort();
+        }
     }
 
 }