+++ src/ninja/mbedded/ninjaterm/view/mainWindow/terminal/txRx/TxRxViewController.java
@@ -20,10 +20,8 @@ import javafx.scene.text.TextFlow;
 import javafx.util.Duration;
 import ninja.mbedded.ninjaterm.model.Model;
 import ninja.mbedded.ninjaterm.model.terminal.Terminal;
-import ninja.mbedded.ninjaterm.util.Decoding.Decoder;
 import ninja.mbedded.ninjaterm.util.streamedText.StreamedText;
 import ninja.mbedded.ninjaterm.util.textNodeInList.TextNodeInList;
-import ninja.mbedded.ninjaterm.view.mainWindow.StatusBar.StatusBarViewController;
 import ninja.mbedded.ninjaterm.view.mainWindow.terminal.txRx.colouriser.ColouriserViewController;
 import ninja.mbedded.ninjaterm.view.mainWindow.terminal.txRx.display.DisplayViewController;
 import ninja.mbedded.ninjaterm.view.mainWindow.terminal.txRx.filters.FiltersViewController;
@@ -369,7 +367,7 @@ public class TxRxViewController {
         //========== ATTACH LISTENER TO LAYOUT =========//
         //==============================================//
 
-        terminal.txRx.display.selectedLayoutOption.addListener((observable, oldValue, newValue) -> {
+        terminal.txRx.display.selLayoutOption.addListener((observable, oldValue, newValue) -> {
             System.out.println("Selected layout option has been changed.");
             updateLayout();
         });
@@ -423,11 +421,28 @@ public class TxRxViewController {
         });*/
     }
 
+    /**
+     * This listener updates the UI with "streamed" RX data. The model is responsible
+     * for calling the listener after RX data has been received and processed.
+     * @param streamedText
+     */
     private void newStreamedTextListener(StreamedText streamedText) {
         ObservableList<Node> observableList = rxDataTextFlow.getChildren();
 
         numCharsInRxTextNodes += streamedText.getText().length();
-        streamedText.shiftToTextNodes(observableList);
+
+        // Move all text/colour info provided in this streamed text object into the
+        // Text nodes that make up the RX data view
+        switch (terminal.txRx.display.selLayoutOption.get()) {
+            case SINGLE_PANE:
+                streamedText.shiftToTextNodes(observableList, observableList.size() - 1);
+                break;
+            case SEPARATE_TX_RX:
+                streamedText.shiftToTextNodes(observableList, observableList.size());
+                break;
+            default:
+                throw new RuntimeException("selLayoutOption not recognised.");
+        }
 
         // Trim RX UI if necessary
         // (the ANSI parser output data is trimmed separately in the model)
@@ -519,8 +534,8 @@ public class TxRxViewController {
      * in the model.
      */
     public void updateLayout() {
-        switch(terminal.txRx.display.selectedLayoutOption.get()) {
-            case COMBINED_TX_RX:
+        switch(terminal.txRx.display.selLayoutOption.get()) {
+            case SINGLE_PANE:
 
                 // Hide TX pane
 
@@ -557,58 +572,11 @@ public class TxRxViewController {
 
                 break;
             default:
-                throw new RuntimeException("selectedLayoutOption unrecognised!");
+                throw new RuntimeException("selLayoutOption unrecognised!");
         }
     }
 
     public void handleKeyTyped(KeyEvent keyEvent) {
-
-        /*if(comPort.isPortOpen() == false) {
-            model.status.addErr("Cannot send data to COM port, port is not open.");
-            return;
-        }
-
-        // Convert pressed key into a ASCII byte.
-        // Hopefully this is only one character!!!
-        byte data = (byte)keyEvent.getCharacter().charAt(0);
-
-        // If to see if we are sending data on "enter", and the "backspace
-        // deletes last typed char" checkbox is ticked
-        if((terminal.txRx.display.selTxCharSendingOption.get() == Display.TxCharSendingOptions.SEND_TX_CHARS_ON_ENTER) &&
-                terminal.txRx.display.backspaceRemovesLastTypedChar.get()) {
-
-            if(keyEvent.getCharacter().equals("\b")) {
-                // We need to remove the last typed char from the "to send" TX buffer
-                terminal.txRx.removeLastCharInTxBuffer();
-                return;
-            }
-
-        }
-
-        // Append the character to the end of the "to send" TX buffer
-        terminal.txRx.addTxCharToSend(data);
-
-        // Check so see what TX mode we are in
-        switch(terminal.txRx.display.selTxCharSendingOption.get()) {
-            case SEND_TX_CHARS_IMMEDIATELY:
-                break;
-            case SEND_TX_CHARS_ON_ENTER:
-                // Check for enter key before sending data
-                if(!keyEvent.getCharacter().equals("\r"))
-                    return;
-                break;
-            default:
-                throw new RuntimeException("selTxCharSendingOption not recognised!");
-        }
-
-        // Send data to COM port, and update stats (both local and global)
-        byte[] dataAsByteArray = fromObservableListToByteArray(terminal.txRx.toSendTxData);
-        comPort.sendData(dataAsByteArray);
-
-        terminal.stats.numCharactersTx.setValue(terminal.stats.numCharactersTx.getValue() + dataAsByteArray.length);
-        model.globalStats.numCharactersTx.setValue(model.globalStats.numCharactersTx.getValue() + dataAsByteArray.length);
-
-        terminal.txRx.txDataSent();*/
 
         terminal.txRx.handleKeyPressed((byte)keyEvent.getCharacter().charAt(0));
     }